refactor: removed utilities folder for util
