+++ src/lib/util/functions.ts
